requirements.txt
streamlit>=1.36.0
pandas
google-genai
requests
google-cloud-firestore